+++ idg-lib/src/CPU/Optimized/kernels/KernelAdderWStack.cpp
@@ -85,9 +85,9 @@ void kernel_adder_wstack(
                         // Add subgrid value to grid
                         for (int pol = 0; pol < NR_POLARIZATIONS; pol++) {
                             int pol_idx = index_pol[pol];
-                            int dst_idx = index_grid(
+                            long dst_idx = index_grid(
                                 NR_POLARIZATIONS, grid_size, w_layer, pol_idx, y_dst, x_dst);
-                            int src_idx = index_subgrid(
+                            long src_idx = index_subgrid(
                                 NR_POLARIZATIONS, subgrid_size, s, pol_idx, y_src, x_src);
                             idg::float2 value = phasor * subgrid[src_idx];
                             value = negative_w ? conj(value) : value;
